+++ src/modules/Language_Switcher/Switchers/Nav.php
@@ -5,9 +5,8 @@
 
 namespace WP_Syntex\Polylang\Language_Switcher\Switchers;
 
-use WP_Syntex\Polylang\Language_Switcher\Elements;
-use WP_Syntex\Polylang\Language_Switcher\Switchers\Types\Nav as Type;
-use WP_Syntex\Polylang\Language_Switcher\Settings\Generic as Settings;
+use PLL_Language;
+use WP_Syntex\Polylang\Language_Switcher\Switchers\Elements\Nav as Element;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -18,18 +17,6 @@ defined( 'ABSPATH' ) || exit;
  */
 class Nav extends Abstract_Switcher {
 	/**
-	 * Constructor.
-	 *
-	 * @since 3.9
-	 *
-	 * @param Settings $settings Instance of `Settings`.
-	 * @param Elements $elements Instance of `Elements`.
-	 */
-	public function __construct( Settings $settings, Elements $elements ) {
-		parent::__construct( $settings, $elements, new Type( $settings ) );
-	}
-
-	/**
 	 * Returns the markup of the switcher.
 	 *
 	 * @since 3.9
@@ -39,14 +26,36 @@ class Nav extends Abstract_Switcher {
 	public function get(): string {
 		$out = '';
 
-		foreach ( $this->elements->get() as $element ) {
-			$out .= $this->item_type->get_row( $element );
+		foreach ( $this->get_elements() as $element ) {
+			$out .= $element->get_row();
 		}
 
 		if ( empty( $out ) || ! $this->settings->show_wrapper ) {
 			return $out;
 		}
 
-		return $this->item_type->wrap( "<ul>{$out}</ul>" );
+		$cr  = $this->settings->preserve_spacing ? "\n" : '';
+		$out = sprintf(
+			'<%1$s id="%2$s" class="%3$s" aria-label="%4$s">%5$s</%1$s>',
+			$this->supports_html5() ? 'nav' : 'div',
+			esc_attr( $this->settings->unique_id ),
+			esc_attr( implode( ' ', $this->get_wrapper_classes() ) ),
+			esc_attr( __( 'Choose a language', 'polylang' ) ),
+			"{$cr}<ul>{$cr}{$out}</ul>"
+		);
+
+		return "{$cr}{$out}{$cr}";
+	}
+
+	/**
+	 * Returns an instance of `Elements\Nav`.
+	 *
+	 * @since 3.9
+	 *
+	 * @param PLL_Language $language Instance of `PLL_Language`.
+	 * @return Element
+	 */
+	protected function get_element( PLL_Language $language ): Element {
+		return new Element( $language, $this->settings );
 	}
 }
